+++ src/components/Header/Header.js
@@ -118,8 +118,8 @@ class Header extends React.Component {
     FeatureFlags.store.listen(this.onFeatureFlagsChange.bind(this));
     // Set the log out link to state
     this.setLogOutLink(window.location.href);
-    // Check if the cookie "PAT_LOGGED_IN" exists and then set the timer to log out the user from
-    // Encore and Catalog(as a side effect)
+    // Set the timer to log out the user from Encore and Catalog
+    // (mainly for Encore while Catalog as a side effect)
     EncoreCatalogLogOutTimer.setEncoreLoggedInTimer(
       window.location.host,
       this.state.currentTime,