+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ========
 ### v2.5.4
-> Update the method to log out users from Encore also check the domain of Catalog as it will log out Catalog as a side effect.
+> Update the method that logs out users from Encore to check the domain of Catalog also since it will log out the users from Catalog at the same time as a side effect.
 
 ### v2.5.3
 > Add the method to completely log out the users by loading the log out endpoint with an invisible iframe.